Step2 Interrupts
Step1 reads the values of the switches and buttons of the Zedboard and
prints them out to the screen using the UART peripheral. The program
lab2.c iterates and actively samples the two GPIO peripherals to read in
their values. This technique, called polling, is not very efficient
because it requires that the CPU spends all if its cycles waiting for
external events to take place. In Step2  the Generic Interrupt
Controller (GIC) of the Zynq7000 PS subsystem is used which manages
interrupts that are coming from different PL and PS sources to the CPU.
It is a centralized resource which is capable of enabling, disabling,
masking and prioritizing interrupt sources sending them to the CPU(s) in
a programmed manner. Once the CPU receives an interrupt from the
interrupt controller, it will save the state of the executing process
and will use the interrupt ID to determine which interrupt handler (aka
interrupt service routine, ISR) to invoke to service the interrupt.